chore: fix server rendering with slots
